normalise anim

+++ Source/AliveLibAO/Animation.hpp
@@ -106,7 +106,7 @@ class Animation final : public AnimationBase
 public:
     virtual void VDecode() override;
 
-    virtual void VRender(s32 xpos, s32 ypos, PrimHeader** ppOt, s16 width, s16 height) override;
+    virtual void VRender(s32 xpos, s32 ypos, PrimHeader** ppOt, s16 width, s32 height) override;
 
     s16 Set_Animation_Data(s32 frameTable, u8** resBlock);
