+++ Source/AliveLibAO/AnimationUnknown.hpp
@@ -13,7 +13,7 @@ class AnimationUnknown final : public AnimationBase
 public:
     virtual void VDecode() override;
 
-    virtual void VRender(s32 xpos, s32 ypos, PrimHeader** ppOt, s16 width, s16 height) override;
+    virtual void VRender(s32 xpos, s32 ypos, PrimHeader** ppOt, s16 width, s32 height) override;
 
     virtual void VCleanUp() override;
